Added readme.md and move the functions draw_pointcloud and register_glfw_callbacks to example.hpp
