+++ examples/software-device/rs-software-device.cpp
@@ -1,18 +1,12 @@
 // License: Apache 2.0. See LICENSE file in root directory.
 // Copyright(c) 2017 Intel Corporation. All Rights Reserved.
-#define M_PI       3.14159265358979323846
+
 #include <librealsense2/rs.hpp> // Include RealSense Cross Platform API
 #include <librealsense2/hpp/rs_internal.hpp>
 
 #include "example.hpp" 
-#include <fstream>
-#include <string>
-#include <math.h>
-
 #include "example.hpp"          // Include short list of convenience functions for rendering
 
-#include <algorithm>            // std::min, std::max
-#include "tclap/CmdLine.h"
 
 #define STB_IMAGE_WRITE_IMPLEMENTATION
 #include <../third-party/stb_image_write.h>
@@ -21,76 +15,7 @@
 #define STB_IMAGE_IMPLEMENTATION
 #include <../third-party/stb_image.h>
 
-using namespace TCLAP;
 using namespace rs2;
-
-struct state {
-    state() : yaw(15.0), pitch(15.0), last_x(0.0), last_y(0.0),
-        ml(false), offset_x(2.f), offset_y(2.f), tex() {}
-    double yaw, pitch, last_x, last_y; bool ml; float offset_x, offset_y; texture tex;
-};
-
-// Handles all the OpenGL calls needed to display the point cloud
-void draw_pointcloud(window& app, state& app_state, rs2::points& points)
-{
-    if (!points)
-        return;
-
-    // OpenGL commands that prep screen for the pointcloud
-    glPopMatrix();
-    glPushAttrib(GL_ALL_ATTRIB_BITS);
-
-    float width = app.width(), height = app.height();
-    //glViewport(0, 0, width, height);
-
-    glClearColor(153.f / 255, 153.f / 255, 153.f / 255, 1);
-    glClear(GL_DEPTH_BUFFER_BIT);
-
-    glMatrixMode(GL_PROJECTION);
-    glPushMatrix();
-    gluPerspective(60, width / height, 0.01f, 10.0f);
-
-    glMatrixMode(GL_MODELVIEW);
-    glPushMatrix();
-    gluLookAt(0, 0, 0, 0, 0, 1, 0, -1, 0);
-
-    glTranslatef(0, 0, +0.5f + app_state.offset_y*0.05f);
-    glRotated(app_state.pitch, 1, 0, 0);
-    glRotated(app_state.yaw, 0, 1, 0);
-    glTranslatef(0, 0, -0.5f);
-
-    glPointSize(width / 640);
-    glEnable(GL_DEPTH_TEST);
-    glEnable(GL_TEXTURE_2D);
-    glBindTexture(GL_TEXTURE_2D, app_state.tex.get_gl_handle());
-    float tex_border_color[] = { 0.8f, 0.8f, 0.8f, 0.8f };
-    glTexParameterfv(GL_TEXTURE_2D, GL_TEXTURE_BORDER_COLOR, tex_border_color);
-    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_S, 0x812F); // GL_CLAMP_TO_EDGE
-    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_T, 0x812F); // GL_CLAMP_TO_EDGE
-    glBegin(GL_POINTS);
-
-
-    /* this segment actually prints the pointcloud */
-    auto vertices = points.get_vertices();              // get vertices
-    auto tex_coords = points.get_texture_coordinates(); // and texture coordinates
-    for (int i = 0; i < points.size(); i++)
-    {
-        if (vertices[i].z)
-        {
-            // upload the point and texture coordinates only for points we have depth data for
-            glVertex3fv(vertices[i]);
-            glTexCoord2fv(tex_coords[i]);
-        }
-    }
-
-    // OpenGL cleanup
-    glEnd();
-    glPopMatrix();
-    glMatrixMode(GL_PROJECTION);
-    glPopMatrix();
-    glPopAttrib();
-    glPushMatrix();
-}
 
 struct synthetic_frame
 {
@@ -184,12 +109,13 @@ int main(int argc, char * argv[]) try
 
 
     window app(1280, 1500, "RealSense Capture Example");
-  
+    state app_state;
+
+    register_glfw_callbacks(app, app_state);
+
     rs2::colorizer color_map; // Save colorized depth for preview
 
     auto last = std::chrono::high_resolution_clock::now();
-
-    state app_state;
 
     float wave_base = 0.f;
     pointcloud pc;
